factory/node-32gb/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
streamlit>=1.40.0
langgraph>=0.2.59
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary]>=3.2.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
